+++ tests/tmc/test_skb_477_xtp_60864.py
@@ -1,0 +1,224 @@
+"""
+This module defines a BDD (Behavior-Driven Development) test scenario
+using pytest-bdd to verify the behavior of the Telescope Monitoring and
+Control (TMC) system resolution of SKB-477.
+"""
+
+
+import json
+
+import pytest
+from assertpy import assert_that
+from pytest_bdd import given, scenario, then, when
+from ska_control_model import ObsState
+from ska_tango_base.commands import ResultCode
+from ska_tango_testing.integration import TangoEventTracer, log_events
+from tango import DevState
+
+from tests.resources.test_harness.central_node_low import CentralNodeWrapperLow
+from tests.resources.test_harness.subarray_node_low import (
+    SubarrayNodeWrapperLow,
+)
+from tests.resources.test_harness.utils.common_utils import JsonFactory
+from tests.resources.test_support.common_utils.tmc_helpers import (
+    prepare_json_args_for_centralnode_commands,
+    prepare_json_args_for_commands,
+)
+from tests.resources.test_support.constant_low import TIMEOUT
+
+
+@pytest.mark.SKA_low
+@scenario(
+    "../features/tmc/SKB_477.feature",
+    "Verify SKB-477 - with TMC entrypoint",
+)
+def test_verify_skb_477():
+    """Test to verify TMC SubarrayNode rejects
+    unnecessary subarray_id key Scan schema
+    """
+
+
+@given("a TMC")
+def given_a_tmc(
+    central_node_low: CentralNodeWrapperLow, event_tracer: TangoEventTracer
+):
+    """
+    This method invokes On command from central node and verifies
+    the state of telescope after the invocation.
+    Args:
+        central_node (CentralNodeWrapperLow): Object of Central node wrapper
+        event_tracer(TangoEventTracer): object of TangoEventTracer used for
+        managing the device events
+    """
+    event_tracer.subscribe_event(
+        central_node_low.central_node, "telescopeState"
+    )
+    event_tracer.subscribe_event(
+        central_node_low.central_node, "longRunningCommandResult"
+    )
+    event_tracer.subscribe_event(central_node_low.subarray_node, "obsState")
+    log_events(
+        {
+            central_node_low.central_node: [
+                "telescopeState",
+                "longRunningCommandResult",
+            ],
+            central_node_low.subarray_node: ["obsState"],
+        }
+    )
+    central_node_low.move_to_on()
+    assert_that(event_tracer).described_as(
+        'FAILED ASSUMPTION IN "GIVEN" STEP: '
+        "'the telescope is is ON state'"
+        "Central Node device"
+        f"({central_node_low.central_node.dev_name()}) "
+        "is expected to be in TelescopeState ON",
+    ).within_timeout(TIMEOUT).has_change_event_occurred(
+        central_node_low.central_node,
+        "telescopeState",
+        DevState.ON,
+    )
+    assert_that(event_tracer).described_as(
+        "FAILED UNEXPECTED INITIAL OBSSTATE: "
+        "Subarray Node device"
+        f"({central_node_low.subarray_node.dev_name()}) "
+        "is expected to be in EMPTY obstate",
+    ).within_timeout(TIMEOUT).has_change_event_occurred(
+        central_node_low.subarray_node,
+        "obsState",
+        ObsState.EMPTY,
+    )
+
+
+@given("Subarray Node in observation state READY")
+def check_subarray_node_obsState_ready(
+    command_input_factory: JsonFactory,
+    central_node_low: CentralNodeWrapperLow,
+    event_tracer: TangoEventTracer,
+    subarray_node_low: SubarrayNodeWrapperLow,
+):
+    """Execute the Configure command with a valid JSON and verify successful
+    execution.
+
+    Args:
+        command_input_factory (JsonFactory): object of TangoEventTracer
+        used for
+        central_node_low (CentralNodeWrapperLow): Object of Central node
+        wrapper
+        event_tracer(TangoEventTracer): object of TangoEventTracer
+        used for managing the device events
+        subarray_node_low (SubarrayNodeWrapperLow): Object of subarray
+        node wrapper
+
+    """
+    event_tracer.subscribe_event(
+        subarray_node_low.subarray_node, "longRunningCommandResult"
+    )
+    log_events({subarray_node_low.subarray_node: ["longRunningCommandResult"]})
+    assign_input_json = prepare_json_args_for_centralnode_commands(
+        "assign_resources_low", command_input_factory
+    )
+    _, unique_id = central_node_low.store_resources(assign_input_json)
+    assert_that(event_tracer).described_as(
+        'FAILED ASSUMPTION IN "GIVEN" STEP: '
+        "'a subarray in READY obsState'"
+        "Subarray Node device"
+        f"({central_node_low.subarray_node.dev_name()}) "
+        "is expected to be in IDLE obstate",
+    ).within_timeout(TIMEOUT).has_change_event_occurred(
+        central_node_low.subarray_node,
+        "obsState",
+        ObsState.IDLE,
+    )
+    assert_that(event_tracer).described_as(
+        'FAILED ASSUMPTION IN "GIVEN" STEP: '
+        "'a subarray in READY obsState'"
+        "Subarray Node device"
+        f"({central_node_low.central_node.dev_name()}) "
+        "is expected have longRunningCommand as"
+        '(unique_id,(ResultCode.OK,"Command Completed"))',
+    ).within_timeout(TIMEOUT).has_change_event_occurred(
+        central_node_low.central_node,
+        "longRunningCommandResult",
+        (unique_id[0], json.dumps((int(ResultCode.OK), "Command Completed"))),
+    )
+    configure_input_json = prepare_json_args_for_commands(
+        "configure_low", command_input_factory
+    )
+    _, unique_id = subarray_node_low.store_configuration_data(
+        configure_input_json
+    )
+    assert_that(event_tracer).described_as(
+        'FAILED ASSUMPTION IN "GIVEN" STEP: '
+        "'a subarray in READY obsState'"
+        "Subarray Node device"
+        f"({central_node_low.subarray_node.dev_name()}) "
+        "is expected to be in READY obstate",
+    ).within_timeout(TIMEOUT).has_change_event_occurred(
+        central_node_low.subarray_node,
+        "obsState",
+        ObsState.READY,
+    )
+    assert_that(event_tracer).described_as(
+        'FAILED ASSUMPTION IN "GIVEN" STEP: '
+        "'a subarray in READY obsState'"
+        "Subarray Node device"
+        f"({central_node_low.subarray_node.dev_name()}) "
+        "is expected have longRunningCommand as"
+        '(unique_id,(ResultCode.OK,"Command Completed"))',
+    ).within_timeout(TIMEOUT).has_change_event_occurred(
+        central_node_low.subarray_node,
+        "longRunningCommandResult",
+        (unique_id[0], json.dumps((int(ResultCode.OK), "Command Completed"))),
+    )
+
+
+@when("I invoke Scan command on Subarray Node with key subarray_id")
+def invoke_scan_with_subarray_id(
+    command_input_factory: JsonFactory,
+    subarray_node_low: SubarrayNodeWrapperLow,
+):
+    """Method to call Scan command using
+      unnecessary subarray_id key in Scan schema
+
+    Args:
+        command_input_factory (JsonFactory): object of TangoEventTracer
+        used for managing the device events
+        subarray_node_low (SubarrayNodeWrapperLow): Object of subarray
+        node wrapper
+    """
+    scan_input_json = prepare_json_args_for_commands(
+        "Scan_low_with_subarray_id_key", command_input_factory
+    )
+    pytest.result, _ = subarray_node_low.subarray_node.Scan(scan_input_json)
+
+
+@then("TMC SubarrayNode raises exception with ResultCode.REJECTED")
+def check_subarray_node_scan_resultcode():
+    """Method to check TMC SubarrayNode returns ResultCode.REJECTED"""
+    assert ResultCode.REJECTED == pytest.result
+
+
+@then("TMC SubarrayNode remains in observation state READY")
+def check_obs_state_ready(
+    event_tracer: TangoEventTracer, subarray_node_low: SubarrayNodeWrapperLow
+):
+    """Method to check observation state of subarray node
+    after configure command.
+
+    Args:
+        event_tracer(TangoEventTracer): object of TangoEventTracer
+        used for managing the device events
+        subarray_node_low (SubarrayNodeWrapperLow): Object of subarray
+        node wrapper
+    """
+    assert_that(event_tracer).described_as(
+        "FAILED UNEXPECTED OBSSTATE: "
+        "Subarray Node device"
+        f"({subarray_node_low.subarray_node.dev_name()}) "
+        "is expected to be in READY obstate",
+    ).within_timeout(TIMEOUT).has_change_event_occurred(
+        subarray_node_low.subarray_node,
+        "obsState",
+        ObsState.READY,
+    )